Remove download's types module
There wasn't much in there, didn't really solve anything that needed solving.
